Move MailmergeError from utils to __init__
