+++ mailmerge/__init__.py
@@ -3,3 +3,7 @@ Mail merge module importable API.
 
 Andrew DeOrio <awdeorio@umich.edu>
 """
+
+
+class MailmergeError(Exception):
+    """Top-level exception raised by mailmerge functions."""
