+++ mailmerge/template_message.py
@@ -14,7 +14,7 @@ import future.backports.email.utils
 import future.backports.email.generator
 import markdown
 import jinja2
-from . import utils
+from . import MailmergeError
 
 # Python 2 pathlib support requires backport
 try:
@@ -62,7 +62,7 @@ class TemplateMessage(object):
         try:
             raw_message = self.template.render(context)
         except jinja2.exceptions.TemplateError as err:
-            raise utils.MailmergeError(
+            raise MailmergeError(
                 "{}: {}".format(self.template_path, err)
             )
         self._message = email.message_from_string(raw_message)
@@ -191,7 +191,7 @@ class TemplateMessage(object):
         """Find attachment file or raise MailmergeError."""
         # Error on empty path
         if not path.strip():
-            raise utils.MailmergeError("Empty attachment header.")
+            raise MailmergeError("Empty attachment header.")
 
         # Create a Path object and handle home directory (tilde ~) notation
         path = Path(path.strip())
@@ -206,7 +206,7 @@ class TemplateMessage(object):
 
         # Check that the attachment exists
         if not path.exists():
-            raise utils.MailmergeError("Attachment not found: {}".format(path))
+            raise MailmergeError("Attachment not found: {}".format(path))
 
         return path
