--- mailmerge/utils.py
@@ -1,9 +1,0 @@
-"""
-Utilities used in multiple mailmerge modules.
-
-Andrew DeOrio <awdeorio@umich.edu>
-"""
-
-
-class MailmergeError(Exception):
-    """Top-level exception raised by mailmerge functions."""
